+++ SourceCode/ChaosRayTracing/ChaosRayTracing/ChaosRayTracing.cpp
@@ -1,3 +1,5 @@
+#include <iostream>
+
 #include "Renderer.hpp"
 
 int main()
