+++ SourceCode/ChaosRayTracing/ChaosRayTracing/Light.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Math3D.hpp"
+
+struct Light
+{
+	float intensity;
+	Vector3 position;
+};
